Update get/set parameters functions
