+++ examples/flowertune-llm/flowertune_llm/models.py
@@ -2,12 +2,17 @@ import math
 
 import torch
 from omegaconf import DictConfig
+from collections import OrderedDict
 from peft import (
     LoraConfig,
     get_peft_model,
+    get_peft_model_state_dict,
+    set_peft_model_state_dict,
 )
 from peft.utils import prepare_model_for_kbit_training
 from transformers import AutoModelForCausalLM, BitsAndBytesConfig
+
+from flwr.common.typing import NDArrays
 
 
 def cosine_annealing(
@@ -56,3 +61,17 @@ def get_model(model_cfg: DictConfig):
     )
 
     return get_peft_model(model, peft_config)
+
+
+def set_parameters(model, parameters: NDArrays) -> None:
+    """Change the parameters of the model using the given ones."""
+    peft_state_dict_keys = get_peft_model_state_dict(model).keys()
+    params_dict = zip(peft_state_dict_keys, parameters)
+    state_dict = OrderedDict({k: torch.Tensor(v) for k, v in params_dict})
+    set_peft_model_state_dict(model, state_dict)
+
+
+def get_parameters(model) -> NDArrays:
+    """Return the parameters of the current net."""
+    state_dict = get_peft_model_state_dict(model)
+    return [val.cpu().numpy() for _, val in state_dict.items()]
